Spatial refactor to carry around flat-field coordinates as well
